# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request a login OTP by email or phone
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RequestOtpInput" }
      responses:
        "200": { description: OTP sent (or dev-mode logged), content: { application/json: { schema: { type: object, properties: { ok: { type: boolean }, dev: { type: boolean }, kind: { type: string, enum: [email, phone] } } } } } }
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify an OTP and start a session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyOtpInput" }
      responses:
        "200": { description: Session token, content: { application/json: { schema: { type: object, properties: { token: { type: string }, id: { type: string }, kind: { type: string }, email: { type: string } } } } } }
...
  /market/crypto/depth:
    get:
      operationId: getCryptoDepth
      tags: [marketdata]
      summary: Real order-book depth (top 20 levels/side) for a crypto symbol, converted to INR
      security: []
      parameters:
        - { name: symbol, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: Order book, content: { application/json: { schema: { $ref: "#/components/schemas/OrderBook" } } } }
...
  /pay/status/{qrId}:
    get:
      operationId: getPayStatus
      tags: [pay]
      summary: Poll whether a QR payment has been credited
      parameters:
        - { name: qrId, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: Status } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token returned by POST /auth/verify-otp.
  schemas:
...
    OrderBookLevel:
      type: object
      properties: { price: { type: number }, qty: { type: number } }
    OrderBook:
      type: object
      properties:
        symbol: { type: string }
        updatedAt: { type: number }
        usdInr: { type: number }
        bids: { type: array, items: { $ref: "#/components/schemas/OrderBookLevel" } }
        asks: { type: array, items: { $ref: "#/components/schemas/OrderBookLevel" } }
...
    RequestOtpInput:
      type: object
      properties: { email: { type: string }, phone: { type: string } }
    VerifyOtpInput:
      type: object
      required: [code]
      properties: { email: { type: string }, phone: { type: string }, code: { type: string } }
...